userAPI 삭제 및 api.js 통합
