+++ src/axios/api.js
@@ -2,34 +2,50 @@ import axios from "axios";
 import { get } from "react-cookie";
 
 const instance = axios.create({
-    baseURL: process.env.REACT_APP_SERVER_URL,
-    headers: {
-        withCredentials: true,
-        "Content-Type": "application/json"
-    }
-})
+  baseURL: `${process.env.REACT_APP_SERVER_URL}/api/`,
+  // withCredentials: true,
+  headers: {
+    withCredentials: true,
+    "Content-Type": "application/json",
+  },
+});
 
-instance.interceptors.request.use((config) => {
-    // request가 보내지기 전에 반영할 요청
-    console.log(get("Authorization"))
-    // console.log("here!:::", useCookies["Authorization"])
-    // if (!!useCookies["Authorization"]) {
-        // const [cookies, setCookie, removeCookie] = useCookies["jwt"]
-        // if (cookies.jwt) {
-        //     config.headers.Authorization = `Bearer ${cookies.jwt}`;
-        // }
-    // }
-
-    return config; 
-}, (error) => { // 에러시의 처리
-    console.log("req err:::", error);
+// 요청 인터셉트
+// 설정된 기능 : 인터셉트하여 console에 찍어줌
+instance.interceptors.request.use(
+  function (config) {
+    console.log("request is ", config);
+    return config;
+  },
+  function (error) {
+    console.log("request error", error);
     return Promise.reject(error);
-})
+  }
+);
 
+// 응답 인터셉트
+// 설정된 기능 : 인터셉트하여 console에 찍어줌
+instance.interceptors.response.use(
+    function (response) {
+      console.log("response is ", response);
+      return response;
+    },
+    function (error) {
+      console.log("response error", error);
+      return Promise.reject(error);
+    }
+  );
+
+export const userAPI = {
+    // user 정보 관련
+    login: (userInfo) => instance.post("/login", userInfo),
+    signUp: (userInfo) => instance.post("/signup", userInfo),
+    logout: () => instance.post("/logout", null)
+}
 
 export const gptAPI = {
-    // layout (nav + main)에서 사용
-    getCredit: async () => await instance.get("/api/credit")
-}
+  // layout (nav + main)에서 사용
+  getCredit: async () => await instance.get("/api/credit"),
+};
 
 export default instance;
